Starting to lock the character actions down.
- We now show a timer that is persistant.
- We also lock down the battle section and crafting section when you are
  adventuring.
- We now displayb a message that the player is dventuring and cannot
  craft or battle.
- Need to implment this for the map section.
- Need to lock a user out of the shop when adventuring.
- Need to lock a player out of the inventory when adventuring.

+++ app/Flare/Models/Monster.php
@@ -45,7 +45,7 @@ class Monster extends Model
         'int'        => 'integer',
         'ac'         => 'integer',
         'gold'       => 'integer',
-        'drop_check' => 'integer',
+        'drop_check' => 'float',
         'max_level'  => 'integer',
     ];
